+++ mlflow/tracing/fluent.py
@@ -8,7 +8,6 @@ import json
 import logging
 from contextvars import ContextVar
 from typing import TYPE_CHECKING, Any, Callable, Generator, Literal, Optional, Union
-from unittest.mock import MagicMock, patch
 
 from cachetools import TTLCache
 from opentelemetry import trace as trace_api
@@ -21,7 +20,6 @@ from mlflow.entities.span_status import SpanStatusCode
 from mlflow.entities.trace_status import TraceStatus
 from mlflow.exceptions import MlflowException
 from mlflow.protos.databricks_pb2 import BAD_REQUEST
-from mlflow.pyfunc import PyFuncModel
 from mlflow.store.tracking import SEARCH_TRACES_DEFAULT_MAX_RESULTS
 from mlflow.tracing import provider
 from mlflow.tracing.constant import (
@@ -174,20 +172,6 @@ def trace(
             return _wrap_function(fn, name, span_type, attributes, model_id)
 
     return decorator(func) if func else decorator
-
-
-def is_traced(model: PyFuncModel):
-    """
-    Check if a PyFuncModel is being traced without logging to the database.
-    """
-
-    with patch("mlflow.tracing.provider._get_tracer", return_value=None) as mock_get_tracer:
-        try:
-            model.predict(MagicMock())
-        except Exception:
-            pass
-
-        return 0 < mock_get_tracer.call_count
 
 
 def _wrap_function(
