+++ mlflow/tracing/utils/__init__.py
@@ -9,6 +9,7 @@ from collections import Counter
 from dataclasses import asdict, is_dataclass
 from functools import lru_cache
 from typing import TYPE_CHECKING, Any, Optional, Union
+from unittest.mock import MagicMock, patch
 
 from opentelemetry import trace as trace_api
 from packaging.version import Version
@@ -457,3 +458,22 @@ def end_client_span_or_trace(
             status=status,
             end_time_ns=end_time_ns,
         )
+
+
+def is_model_traced(model):
+    """
+    Check if a PyFuncModel is being traced without logging to the database.
+
+    Args:
+        model (PyFuncModel): The model to check.
+
+    Returns:
+        True if the model is being traced, False otherwise.
+    """
+    with patch("mlflow.tracing.provider._get_tracer", return_value=None) as mock_get_tracer:
+        try:
+            model.predict(MagicMock())
+        except Exception:
+            pass
+
+        return 0 < mock_get_tracer.call_count
